# Agentic_outputs/5jun/phase_3_domain_design/api_contracts.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCMS API",
    "version": "1.0.0",
    "description": "Generic Docker Container Management System REST API. Provides full lifecycle management of containers, images, networks, volumes, clusters, monitoring, logging, and user administration across multi-host Docker Swarm deployments."
  },
  "servers": [
    {
      "url": "https://api.dcms.internal/v1",
      "description": "Production API Gateway"
    },
    {
      "url": "https://api.staging.dcms.internal/v1",
      "description": "Staging API Gateway"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authenticate user and obtain JWT tokens",
        "tags": ["auth"],
        "security": [],
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "admin@example.com" },
                  "password": { "type": "string", "format": "password", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string", "description": "JWT access token" },
                    "refresh_token": { "type": "string", "description": "Opaque refresh token" },
                    "expires_in": { "type": "integer", "description": "Access token TTL in seconds", "example": 3600 },
                    "token_type": { "type": "string", "example": "Bearer" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Obtain a new access token using a refresh token",
        "tags": ["auth"],
        "security": [],
        "operationId": "authRefresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "refresh_token": { "type": "string" },
                    "expires_in": { "type": "integer" },
                    "token_type": { "type": "string", "example": "Bearer" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke the current session and invalidate refresh token",
        "tags": ["auth"],
        "operationId": "authLogout",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refresh_token"],
                "properties": {
                  "refresh_token": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Logged out successfully" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "Malformed request or validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "RS256-signed JWT issued by /auth/login. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned for all 4xx and 5xx responses",
        "required": ["error", "message"],
        "properties": {
          "error": { "type": "string", "description": "Machine-readable error code", "example": "CONTAINER_NOT_FOUND" },
          "message": { "type": "string", "description": "Human-readable error description" },
          "details": {
            "type": "array",
            "description": "Field-level validation errors",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "issue": { "type": "string" }
              }
            }
          },
          "trace_id": { "type": "string", "description": "Distributed trace ID for log correlation" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}